Change all back end representations of progress components to large numerics

+++ app/graphql/types/util/Positive.scala
@@ -7,12 +7,12 @@ import sangria.macros.derive.{ InputObjectTypeName, deriveInputObjectType, deriv
 import sangria.schema.{ InputObjectType, ObjectType }
 
 sealed abstract case class Positive(
-    positive: Int
+    positive: BigInt
 )
 
 object Positive {
 
-  private def apply(positive: Int): Positive = new Positive(positive) {}
+  private def apply(positive: BigInt): Positive = new Positive(positive) {}
 
   implicit val positiveCodec: Codec[Positive] =
     deriveCodec[Positive].iemap(p => Right(p).filterOrElse(_.positive > 0, "Not a positive natural number"))(identity)
